backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary>=2.9.11
python-dotenv==1.0.0
flask-bcrypt==1.0.1
requests==2.31.0
email-validator==2.1.0
